requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.2
attrs==22.2.0
azure-common==1.1.28
azure-core==1.29.7
azure-identity==1.15.0
azure-search-documents==11.4.0b8
boto3==1.34.23
botocore==1.34.23
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
cryptography==41.0.7
dataclasses-json==0.6.3
exceptiongroup==1.2.0
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langsmith==0.0.83
marshmallow==3.20.2
msal==1.26.0
msal-extensions==1.1.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==0.27.4
packaging==23.2
portalocker==2.8.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.28.2
s3transfer==0.10.0
six==1.16.0
slack-bolt==1.17.0
slack-sdk==3.20.2
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.15
yarl==1.8.2
